got it to thread with 4 threads
