+++ R/RcppExports.R
@@ -5,6 +5,10 @@ mediate_helper <- function(env) {
     invisible(.Call(`_reverseC_mediate_helper`, env))
 }
 
+threaded_mediate_helper <- function(env) {
+    invisible(.Call(`_reverseC_threaded_mediate_helper`, env))
+}
+
 cpp_mult <- function(m1, m2) {
     .Call(`_reverseC_cpp_mult`, m1, m2)
 }
