+++ src/double_matrix_indexed_vector_ref.cpp
@@ -1,0 +1,3 @@
+#include "double_matrix_indexed_vector_ref.hpp"
+
+// [[Rcpp::plugins(cpp11)]]
