+++ src/double_matrix_indexed_vector_ref.hpp
@@ -1,0 +1,14 @@
+#ifndef CLASS_DOUBLEMATRIX_INDEXED_VECTOR_REF_HPP
+#define CLASS_DOUBLEMATRIX_INDEXED_VECTOR_REF_HPP
+#include <Rcpp.h>
+
+// [[Rcpp::plugins(cpp11)]]
+
+class DoubleMatrix_Indexed_Vector_Ref;
+#include "double_matrix.hpp"
+
+class DoubleMatrix_Indexed_Vector_Ref{
+  
+};
+
+#endif //CLASS_DOUBLEMATRIX_INDEXED_VECTOR_REF_HPP
