+++ test_commands.txt
@@ -1,17 +1,29 @@
 devtools::load_all()
-param = reverseC::MediateDataGenerationParameters(nSim=1)
+
+param = reverseC::MediateDataGenerationParameters(n=1000, nSim=1)
 data_item = reverseC::generateDataMatrix(param)[[1]]
 models = reverseC::assembleLinearModels(data_item)
-Rcpp::sourceCpp('src/mediation_helper_functions.cpp')
-source('R/mediate_with_rcpp.R')
-med_env = stripped_down_mediate_with_rcpp(models@med.fit, models@out.fit, treat="X", mediator="M1", return_context=T)
-export_environment(med_env)
+med_env = reverseC::stripped_down_mediate_with_rcpp(models@med.fit, models@out.fit, treat="X", mediator="M1", return_context=T)
+reverseC::export_environment(med_env)
 
 
-tic("test");stripped_down_mediate_with_rcpp(models@med.fit, models@out.fit, treat="X", mediator="M1");toc()
+
+reverseC::mediate_helper(environment())
+
+
+reverseC::threaded_mediate_helper(environment())
+
+
+stripped_down_mediate_with_rcpp(models@med.fit, models@out.fit, treat="X", mediator="M1")
+
 
 
 df = data.frame("N" = c(1.1,1.2,1.3))
 df["A"] = c(1,2,3)
 df["B"] = c(4,5,6)
 dft = as.data.frame(t(df))
+
+
+mmat = matrix(c(1,2,3,4,5,6,7,8,9,10,11,12), nrow=4, ncol=3)
+tmat = t(mmat)
+reverseC::cpp_mult(mmat,tmat)
